refactor / split out collection related components
