+++ frontend/src/metabase/collections/constants.js
@@ -1,0 +1,2 @@
+// unit to use for calculating collection sidebar spacing (in px)
+export const SIDEBAR_SPACER = 14
